Refactor: Move FileCombineCommand to Files namespace and add FileEncodingHelper for encoding detection
